backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
passlib[bcrypt]
python-jose
openai
